added server_compare_charts component
